pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuble"
version = "1.0.0"
description = "NUBLE - Institutional-Grade AI Investment Research Platform"
authors = [{name = "NUBLE"}]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "rich",
    "inquirer", 
    "tiktoken",
    "requests",
    "openai",
    "anthropic",
    "python-dotenv",
    "numpy"
]

[project.scripts]
nuble = "nuble.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"